Customer Dao related changes

+++ src/main/java/com/javasuitor/brokerpoint/dao/CustomerDao.java
@@ -6,10 +6,11 @@ import com.javasuitor.brokerpoint.entity.Customer;
 
 public interface CustomerDao {
 	
-	public Customer get(Long customerId);
+	public Customer get(Integer customerId);
 	public Customer save(Customer customer);
-	public Customer update(Long customerId , Customer customer);
-	public Long delete(Long customerId);	
+	public Customer update(Integer customerId , Customer customer);
+	public Integer delete(Integer customerId);	
 	public List<Customer> getAllCustomers();
+	
 
 }
