+++ src/main/java/com/javasuitor/brokerpoint/service/CustomerService.java
@@ -5,11 +5,11 @@ import java.util.List;
 import com.javasuitor.brokerpoint.entity.Customer;
 
 public interface CustomerService {
-	
-	public Customer get(Long customerId);
+	public Customer get(Integer customerId);
 	public Customer save(Customer customer);
-	public Customer update(Long customerId , Customer customer);
-	public Long delete(Long customerId);
+	public Customer update(Integer customerId , Customer customer);
+	public Integer delete(Integer customerId);
 	public List<Customer> getAllCustomers();
+	
 
 }
